refactored cache operation
